# GH-45201: [C++][Parquet] Improve performance of generating size statistics