--- a/cpp/src/parquet/size_statistics.h
+++ b/cpp/src/parquet/size_statistics.h
@@ -17,9 +17,11 @@
 
 #pragma once
 
+#include <cstdint>
 #include <optional>
 #include <vector>
 
+#include "arrow/util/span.h"
 #include "parquet/platform.h"
 #include "parquet/type_fwd.h"
 
@@ -88,5 +90,9 @@ struct PARQUET_EXPORT SizeStatistics {
   /// \brief Make an empty SizeStatistics object for specific type.
   static std::unique_ptr<SizeStatistics> Make(const ColumnDescriptor* descr);
 };
+
+PARQUET_EXPORT
+void UpdateLevelHistogram(::arrow::util::span<const int16_t> levels,
+                          ::arrow::util::span<int64_t> histogram);
 
 }  // namespace parquet