@pitrou 2025-01-09
Yes, it's even better, see the new numbers in the PR description.