@pitrou 2025-01-09
The Valgrind error is unrelated.